fix: full refactor old code

+++ src/main/java/dev/zhulidov/cash_tracker/service/CategoryService.java
@@ -8,6 +8,7 @@ import dev.zhulidov.cash_tracker.repository.CategoryRepository;
 import dev.zhulidov.cash_tracker.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import dev.zhulidov.cash_tracker.util.SecurityUtils;
 
 import java.util.List;
 
@@ -17,9 +18,9 @@ public class CategoryService {
   private final   CategoryRepository repository;
   private final   UserRepository userRepository;
 
-    public CategoryDto createCategory(CategoryCreateRequestDto requestDto){
-        var user = userRepository.findById(requestDto.userId()).orElseThrow(
-                ()-> new ResourceNotFoundException("User with id: "+requestDto.userId()+" not found")
+    public CategoryDto createCategory(CategoryCreateRequestDto requestDto, Long userId){
+        var user = userRepository.findById(userId).orElseThrow(
+                ()-> new ResourceNotFoundException("User with id: "+userId+" not found")
         );
         var category = Category.builder()
                 .categoryName(requestDto.categoryName())
@@ -29,20 +30,27 @@ public class CategoryService {
         return new CategoryDto(savedCategory.getId(), savedCategory.getCategoryName(),savedCategory.getUser().getName());
     }
 
-    public CategoryDto getCategoryById(Long id){
+    public CategoryDto getCategoryById(Long id, Long userId){
+
         var category = repository.findById(id).orElseThrow(()-> new ResourceNotFoundException("Category not found"));
+        SecurityUtils.assertOwner(category.getUser().getId(),userId);
         return new CategoryDto(category.getId(), category.getCategoryName(), category.getUser().getName());
     }
 
-    public void deleteCategoryById(Long id){
+    public void deleteCategoryById(Long id, Long userId){
+        var category = repository.findById(id).orElseThrow(()-> new ResourceNotFoundException("Category not found"));
+        SecurityUtils.assertOwner(category.getUser().getId(),userId);
         repository.deleteById(id);
+
     }
 
-    public CategoryDto updateCategory(Long id, String categoryName){
+    public CategoryDto updateCategory(Long id, String categoryName, Long userId){
        Category category = repository.findById(id).orElseThrow(()-> new ResourceNotFoundException("Category not found"));
+        SecurityUtils.assertOwner(category.getUser().getId(),userId);
         category.setCategoryName(categoryName);
         var savedCat = repository.save(category);
         return new CategoryDto(savedCat.getId(),savedCat.getCategoryName(), savedCat.getUser().getName());
+
     }
 
     public List<CategoryDto> getCategoriesByUserId(Long userId){
